+++ example/house/house_greebled.py
@@ -1,12 +1,14 @@
 import cadquery as cq
-from cqfantasy.house import House, TudorSplitBody, ShingleRoof
+from cqfantasy.house import House, TudorSplitBody, ShingleRoof, WallTudor
 
 bp_house = House()
 bp_house.length = 100
-bp_house.width = 180
+bp_house.width = 175
 bp_house.height = 75
 bp_house.roof_height = 50
 bp_house.roof_overhang = (15,10)
+bp_house.tile_height = 1.5
+bp_house.floor_height = 5
 
 bp_house.render_roof = True
 bp_house.render_doors = True
@@ -15,10 +17,7 @@ bp_house.render_doors = True
 #bp_house.bp_body = TudorBody()
 #bp_house.bp_body = StuccoBrickBody()
 bp_house.bp_body = TudorSplitBody()
-
-bp_house.floor_height = 5
 bp_body = bp_house.bp_body
-
 bp_body.wall_width = 4
 bp_body.split_divide_height = 25
 bp_body.x_styles = ["right",None,None,'left']
@@ -29,12 +28,16 @@ bp_door.height = 50
 bp_door.width = 2.5
 bp_house.door_cut_width_padding = 10
 
-bp_house.tile_height = 1.5
-
 bp_house.bp_roof = ShingleRoof()
 bp_roof = bp_house.bp_roof
 bp_roof.tile_length = 10
 bp_roof.overhang_inset = (12,10,4)
+
+bp_roof.bp_outside_wall = WallTudor()
+bp_roof.bp_outside_wall.panel_length = 130/4
+bp_roof.bp_outside_wall.panel_space:float = 2
+bp_roof.bp_outside_wall.panel_width:float = 2.5 
+bp_roof.bp_outside_wall.styles = "cross"
 
 bp_house.window_x_style = [None,None,None]
 bp_house.window_y_style = [None,'window',None,'window',None,'window',None]