+++ example_runner.py
@@ -16,6 +16,7 @@ import example.house.body_types
 import example.house.house_greebled
 import example.house.tile_generator
 import example.house.shingle_roof
+import example.house.WallTudor
 
 #------------------------------
 ## Wall Examples
@@ -48,7 +49,7 @@ import example.tower.round_block_stucco_generator
 
 
 #------------------------------
-run_base = False
+run_base = True
 
 if run_base:
     print('Tower Examples')
@@ -85,7 +86,7 @@ if run_base:
     print('end tower_top')
 
 #------------------------------
-run_greebled = False
+run_greebled = True
 
 if run_greebled:
     print('start tower_greebled_example')
@@ -124,7 +125,7 @@ if run_greebled:
 
 
 #-----------------------------------
-run_alt = False
+run_alt = True
 
 if run_alt:
     print('start tower_alt_greebled_example')
@@ -136,7 +137,7 @@ if run_alt:
     print('end tower_alt_greebled_example')
 
 #-----------------------------------
-run_stucco = False
+run_stucco = True
 
 if run_stucco:
     print('start tower_stucco_example')